combine base and implementation documentation
